chore: file renames to match the *_ext / *_handle / *_store conventions
- graph/{retry,timeout,petgraph}.rs → graph/{retry_ext,timeout_ext,petgraph_ext}.rs
  (the files contain extension traits/impls; the `_ext` suffix matches
  the trait names RetryExt / TimeoutExt and reads better at import
  sites.)
- registry/{cache,content,key,store}.rs →
  registry/{resource_cache,content_handle,composite_key,registry_store}.rs
  (each file's primary type is the noun in the new name; the prefixed
  variants disambiguate from same-named types elsewhere.)
- detection/pattern_engine.rs: document the two PatternEngineRef
  variants (Shared = process-wide singleton via
  PatternEngine::instance, Owned = freshly compiled per-config
  engine). Manual Deref impl stays — derive_more::Deref doesn't
  support enums.
