+++ crates/nvisy-engine/src/graph/mod.rs
@@ -7,11 +7,11 @@
 //! - [`RetryExt`]: automatic retry with configurable backoff.
 //! - [`TimeoutExt`]: wall-clock deadline enforcement for pipeline phases.
 
-mod petgraph;
-mod retry;
-mod timeout;
+mod petgraph_ext;
+mod retry_ext;
+mod timeout_ext;
 
-pub(crate) use self::petgraph::GraphExt;
+pub(crate) use self::petgraph_ext::GraphExt;
 #[allow(unused_imports)] // wired when operations gain internal retry
-pub(crate) use self::retry::RetryExt;
-pub(crate) use self::timeout::TimeoutExt;
+pub(crate) use self::retry_ext::RetryExt;
+pub(crate) use self::timeout_ext::TimeoutExt;
