+++ crates/nvisy-engine/src/graph/petgraph_ext.rs
