+++ crates/nvisy-engine/src/registry/fjall_ext.rs
@@ -11,7 +11,7 @@ use fjall::{Database, Keyspace, KeyspaceCreateOptions, KvSeparationOptions};
 use nvisy_core::{Error, ErrorKind, Result};
 use uuid::Uuid;
 
-use super::key::CompositeKey;
+use super::composite_key::CompositeKey;
 
 const COMPONENT: &str = "registry";
