+++ crates/nvisy-engine/src/registry/resource_cache.rs
